requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
boto3==1.34.36
python-dotenv==1.0.0
pandas==1.5.3
databricks-sql-connector==2.1.0
